pyproject.toml
[project]
name = "what-does-transformer-look-at"
version = "0.1.0"
description = "Add a short description here"
authors = [
    { name = "mizuki@kanolab", email = "trombonelover8888@gmail.com" }
]
dependencies = ["pyyaml~=6.0", "matplotlib~=3.7.1", "seaborn~=0.12.2", "torch~=2.0.1", "transformers~=4.30.2", "ipadic~=1.0.0", "fugashi~=1.2.1", "japanize_matplotlib~=1.1.3"]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
[tool.hatch.metadata]
allow-direct-references = true
